Fix OsError creation
